prints all the elements of a listint_t list

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -1,5 +1,4 @@
 #include <stdio.h>
-#include <string.h>
 #include "lists.h"
 
 /**
@@ -11,7 +10,7 @@
 size_t print_listint(const listint_t *h)
 {
 	const listint_t *current = h;
-	size_t size = 0;
+	int size = 0;
 
 	while (current != NULL)
 	{
